[test] usercoupon 추가, coupon 수정
