+++ src/main/java/com/fouribnb/coupon/domain/entity/Coupon.java
@@ -4,6 +4,7 @@ import com.fouribnb.coupon.presentation.dto.request.GrantCouponRequestDto;
 import com.fouribnb.coupon.presentation.dto.request.UpdateCouponRequestDto;
 import com.fourirbnb.common.domain.BaseEntity;
 import com.fourirbnb.common.exception.OperationNotAllowedException;
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -11,6 +12,7 @@ import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import java.util.UUID;
 import lombok.Builder;
@@ -33,34 +35,16 @@ public class Coupon extends BaseEntity {
     @Column(nullable = false, name = "coupon_id")
     private UUID id;
 
-    @Column(name = "user_id")
-    private Long userId;
-
-    @Column( name = "payment_id")
-    private UUID paymentId;
-
     @Column(nullable = false, name = "coupon_name")
     private String couponName;
-
-    @Column(nullable = false, name = "coupon_status")
-    @Enumerated(EnumType.STRING)
-    private CouponStatus couponStatus;
 
     @Column(nullable = false, name = "discount_value")
     private Long discountValue;
 
-    @Column(nullable = false, name = "is_used")
-    private boolean isUsed;
-
     @Builder
-    public Coupon(Long userId, UUID paymentId, String couponName, CouponStatus couponStatus,
-            Long discountValue, boolean isUsed) {
-        this.userId = userId;
-        this.paymentId = paymentId;
+    public Coupon(String couponName, Long discountValue) {
         this.couponName = couponName;
-        this.couponStatus = couponStatus;
         this.discountValue = discountValue;
-        this.isUsed = isUsed;
     }
 
     public void update(UpdateCouponRequestDto dto){
@@ -72,16 +56,6 @@ public class Coupon extends BaseEntity {
         }
     }
 
-    public void grant(GrantCouponRequestDto dto) {
-        if(this.couponStatus == CouponStatus.valueOf("ACTIVE")){
-            if(this.isUsed == false){
-                this.userId = dto.getUserId();
-            }else {
-                throw new OperationNotAllowedException("이미 사용된 쿠폰입니다");
-            }
-        }else {
-            throw new OperationNotAllowedException("만료된 쿠폰입니다");
-        }
-    }
+
 }
